+++ trailblazer/store/core.py
@@ -1,10 +1,12 @@
+from trailblazer.store.crud.create import CreateHandler
 from trailblazer.store.crud.read import ReadHandler
 from trailblazer.store.crud.update import UpdateHandler
 
 
 class CoreHandler(
-    UpdateHandler,
+    CreateHandler,
     ReadHandler,
+    UpdateHandler,
 ):
     """Aggregating class for the store API handlers."""
